hasnt crushed so far
